Multiple bug fixes
- Fixed volume and sensitivity sliders not saving
- Fixed bug to get out of bounds on level 1
- Fixed what should be most texture issues on level 2
- Fixed ability to unpause menu on options screen leaving options screen visible
- Fixed ability to shoot while paused/while menu was up
- Fixed the player being able to shoot with the control key
- Fixed hit marker not being aligned with crosshair
- Fixed reload prompt not disappearing after reloading

+++ Assets/Scripts/playerSettings.cs
@@ -8,5 +8,5 @@ public class playerSettings : ScriptableObject
 {
     public int horizontalSensitivity;
     public int verticalSensitivity;
-    public int volumePercent;
+    public float volumePercent;
 }
